Tested incorporation of rect class into game for overlaps
Seems like there is some error where the rectangles remain on screen before disappearing. Also an error in the over laps pushback function as it is probably not being updated properly and is just causing the rectangles to jump somewhere instead of actually colliding with one another. Will revisit after creating some more logic to the game and updating the game class